Chanfed some folder casing / moved some files around
