+++ src/app/features/Loadouts/Loadout/WeaponList/Weapon/AttachmentList/LoadoutWeaponAttachmentList.js
@@ -1,7 +1,7 @@
 import React, { useState, useContext, useCallback } from 'react'
 import PropTypes from 'prop-types'
 
-import { LoadoutContext } from 'app/features/Loadouts'
+import { LoadoutContext } from 'app/features/loadouts'
 import AddArmoryItemDialog from 'app/shared/components/Dialogs/AddArmoryItemDialog'
 import AddButton from 'app/shared/components/Buttons/AddButton'
 import LoadoutWeaponAttachment from './Attachment/LoadoutWeaponAttachment'
