+++ src/app/features/Share/SharedLoadout.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { Loading, Error } from 'app/shared/components'
 import ReactiveTitle from 'app/shared/components/Text/ReactiveTitle'
-import { LoadoutView } from 'app/features/Loadouts'
+import { LoadoutView } from 'app/features/loadouts'
 import SharedNotFound from './SharedNotFound'
 import database from '../../../firebase/database'
