+++ src/app/features/events/EventList/Calendar/index.js
@@ -1,0 +1,3 @@
+export { default as CalendarToolbar } from './CalendarToolbar'
+export { default as CalendarEvent } from './CalendarEvent'
+export { default as CalendarAgendaEvent } from './CalendarAgendaEvent'
