+++ src/app/shared/components/Dialogs/ConfirmDeleteDialog.js
